Added confirmation dialog when deleting list

+++ src/pages/Lists/index.js
@@ -1,5 +1,6 @@
 import React, { PureComponent } from 'react';
 import Helmet from 'react-helmet';
+import confirm from 'reactstrap-confirm';
 import { connect } from 'react-redux';
 import { orderBy } from 'lodash';
 import { toastAlert } from '../../utils/helpers';
@@ -12,7 +13,6 @@ import AddSiteButton from '../../components/AddSiteButton';
 class Lists extends PureComponent {
   constructor() {
     super();
-
     this.addNewList = this.addNewList.bind(this);
     this.onDeleteWebsite = this.onDeleteWebsite.bind(this);
   }
@@ -21,17 +21,26 @@ class Lists extends PureComponent {
     this.props.list();
   }
 
-  onDeleteWebsite = id => {
-    this.props.deleteList(id).then(res => {
-      if (res.type === 'LIST_DELETE_FAIL') {
-        toastAlert('error', `Something went wrong deleting the list`);
-      } else {
-        toastAlert('info', `Successfully deleted the list`);
-      }
+  onDeleteWebsite = async (id, name) => {
+    const confirmed = await confirm({
+      title: `Deleting list "${name}"`,
+      message: 'Are you sure you want to delete this list?',
+      confirmText: `Yes, I'm sure!`,
+      confirmColor: 'primary',
+      cancelColor: 'link text-danger'
     });
+    if (confirmed) {
+      this.props.deleteList(id).then(res => {
+        if (res.type === 'LIST_DELETE_FAIL') {
+          toastAlert('error', `Something went wrong deleting the list`);
+        } else {
+          toastAlert('info', `Successfully deleted the list`);
+        }
+      });
+    }
   };
 
-  addNewList() {
+  addNewList = async () => {
     this.props.create().then(res => {
       if (res.type === 'CREATE_LIST_REQUEST_FAIL') {
         toastAlert('error', `Something went wrong adding a new list`);
@@ -40,7 +49,7 @@ class Lists extends PureComponent {
         this.props.history.push(`/list/${res.newListId.data._id}`);
       }
     });
-  }
+  };
 
   render() {
     const { isLoading, data } = this.props.lists;
@@ -50,7 +59,7 @@ class Lists extends PureComponent {
         <List
           key={list._id}
           list={list}
-          onDeleteWebsite={() => this.onDeleteWebsite(list._id)}
+          onDeleteWebsite={() => this.onDeleteWebsite(list._id, list.name)}
         />
       ));
     } else {
